--- src/test/java/io/aiven/kafka/auth/AivenKafkaPrincipalBuilderTest.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2021 Aiven Oy https://aiven.io
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.aiven.kafka.auth;
-
-import io.aiven.kafka.auth.utils.TimeWithTimer;
-
-public class AivenKafkaPrincipalBuilderTest extends AivenKafkaPrincipalBuilderTestBase {
-    @Override
-    protected AivenKafkaPrincipalBuilder createBuilder(final TimeWithTimer time) {
-        return new AivenKafkaPrincipalBuilder(time);
-    }
-}
